[fix] Change revoke code for OWNERSHIP privilege since it was dropped from Snowflake (#360)
As discussed in this issue (https://github.com/chanzuckerberg/terraform-provider-snowflake/issues/304), since 10/2020 Snowflake dropped support for OWNERSHIP revoking.
The proposed solution will use CURRENT_ROLE() to return Terraform's role and grant the OWNERSHIP to it. This way we are still able to grant OWNERSHIP from the resource to any other role via Terraform.