[breaking] Fix to properly distinguish between similarly named views in other databases and schemas (#413)
<!-- Feel free to delete comments as you fill this in -->

<!-- summary of changes -->
If I had a view `foo.bar.view` and another view `foo.baz.view`, the provider was not able to distinguish between them. This caused all sort of issues such as modifying the wrong view, reassigning grants, etc.

NOTE: It seems to me that we should require a DB and a SCHEMA to be specified in the resource. The lack of one or the other makes it so that the queries are "context dependent", under-specified, and somewhat dangerous. I decided to make both db and schema required fields and simplify the code accordingly. LMK what your thoughts are.

## Test Plan
<!-- detail ways in which this PR has been tested or needs to be tested -->
* [X] acceptance tests
<!-- add more below if you think they are relevant -->
* [X] Unit tests

## References
<!-- issues documentation links, etc  -->

* https://docs.snowflake.com/en/sql-reference/sql/alter-view.html
* https://docs.snowflake.com/en/sql-reference/sql/show-views.html