fix: Stream append only (#653)
This change updates the logic used when reading the data for the `snowflake_stream` resource to determine whether or not the stream is append only.

Per the documentation at <https://docs.snowflake.com/en/sql-reference/sql/show-streams.html>, we can determine whether a stream is append only by checking the value of the `mode` for the stream. If the value is `APPEND_ONLY`, then the stream is append only.

## Test Plan
* [X] Unit Tests
* [x] Acceptance Tests

If possible I would like feedback on the acceptance tests. I'm not terribly familiar with terraform provider testing.

## References

* Related issue: <https://github.com/chanzuckerberg/terraform-provider-snowflake/issues/652>